requirements.txt
langchain>=0.0.267
langchain-community>=0.0.16
python-dotenv>=1.0.0
requests>=2.31.0 